Move VersionParser under the TestFramework namespace
