+++ src/TestFramework/Codeception/Adapter/CodeceptionAdapterFactory.php
@@ -40,13 +40,13 @@ use Infection\TestFramework\Coverage\XmlReport\JUnitTestCaseSorter;
 use Infection\TestFramework\TestFrameworkAdapter;
 use Infection\TestFramework\TestFrameworkAdapterFactory;
 use Infection\TestFramework\TestFrameworkConfigParseException;
-use Infection\Utils\VersionParser;
+use Infection\TestFramework\VersionParser;
 use LogicException;
-use function Safe\file_get_contents;
-use function Safe\sprintf;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Yaml\Exception\ParseException;
 use Symfony\Component\Yaml\Yaml;
+use function Safe\file_get_contents;
+use function Safe\sprintf;
 
 /**
  * @internal
