+++ src/TestFramework/PhpUnit/Adapter/PhpUnitAdapterFactory.php
@@ -44,9 +44,9 @@ use Infection\TestFramework\PhpUnit\Config\Path\PathReplacer;
 use Infection\TestFramework\PhpUnit\Config\XmlConfigurationHelper;
 use Infection\TestFramework\TestFrameworkAdapter;
 use Infection\TestFramework\TestFrameworkAdapterFactory;
-use Infection\Utils\VersionParser;
-use function Safe\file_get_contents;
+use Infection\TestFramework\VersionParser;
 use Symfony\Component\Filesystem\Filesystem;
+use function Safe\file_get_contents;
 
 /**
  * @internal
