+++ src/TestFramework/VersionParser.php
@@ -33,11 +33,11 @@
 
 declare(strict_types=1);
 
-namespace Infection\Utils;
+namespace Infection\TestFramework;
 
+use Webmozart\Assert\Assert;
 use function Safe\preg_match;
 use function Safe\sprintf;
-use Webmozart\Assert\Assert;
 
 /**
  * @internal
